* tag 'build-30':
  Build 30
  Fix for Interface Builder
  Add synchronization of keyboard layouts and input sources
  Fix case
  Fix shift gestures in automatic character script conversion
  Refactor
  Refactor
  Fix automatic script conversion after combining marks
  Fix status item
  Refactor
  Add selector keys for character search window
  Remove appending of script code item to frequently used characters
  Add character search by short script code
  Refactor
  Fix input in numeric fields
  Add automatic character conversion to script of previous letter
  Fix deployment target for macOS unit tests
  Refactor the mechanism for changing appearance

+++ Keyboard/Keyboard.swift
@@ -13,6 +13,10 @@ class Keyboard: NSObject {
     static let `default`: Keyboard = .init()
     var delegate: KeyboardDelegate?
     
+    var scriptComponent: CharacterComponent? {
+        return delegate?.documentContext.beforeInput?.applyingTransform(.stripCombiningMarks, reverse: false)?.trimmingCharacters(in: CharacterSet.letters.inverted).last?.characterComponents.filter {CharacterComponent.scripts.contains($0)}.first
+    }
+    
     var version: String {
         let versionNumber = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") as! String
         let buildNumber = Bundle.main.object(forInfoDictionaryKey: "CFBundleVersion") as! String
@@ -146,10 +150,14 @@ class Keyboard: NSObject {
     
     private var shouldDeletePreviousCharacter: Bool = false
     
+    private var previousDocumentContextBeforeInput: String = .init()
+    
     private func input() {
         guard let currentKey = currentKey else {
             return
         }
+        
+        previousDocumentContextBeforeInput = delegate?.documentContext.beforeInput ?? .init()
         
         for shiftFlagDirection in shiftFlagDirections {
             switch shiftFlagDirection {
@@ -301,12 +309,14 @@ class Keyboard: NSObject {
             
             shouldDeletePreviousCharacter = true
             
-            let mixingComponents = characterComponents.map {CharacterComponent.letterToMixingComponentDictionary[$0] ?? $0}
-            let combiningComponents = characterComponents.map {CharacterComponent.letterToCombiningComponentDictionary[$0] ?? $0}
+            let modifierCharacterComponents = characterComponents.removing(characterComponents: CharacterComponent.scripts)
+            
+            let mixingComponents = modifierCharacterComponents.map {CharacterComponent.letterToMixingComponentDictionary[$0] ?? $0}
+            let combiningComponents = modifierCharacterComponents.map {CharacterComponent.letterToCombiningComponentDictionary[$0] ?? $0}
             
             let combiningSuffix: [CharacterComponent] = [direction == .left ? .combining : (direction == .upLeft ? .above : .below)]
             
-            var combiningCharacter: String = (characterComponents + combiningSuffix).character
+            var combiningCharacter: String = (modifierCharacterComponents + combiningSuffix).character
             combiningCharacter = (combiningComponents + combiningSuffix).character
             
             guard combiningCharacter.isEmpty else {
@@ -314,7 +324,7 @@ class Keyboard: NSObject {
                 break
             }
             
-            let ligatureCharacter = (previousCharacter.characterComponents + characterComponents).character
+            let ligatureCharacter = (previousCharacter.characterComponents + modifierCharacterComponents).character
             let combinedCharacter = (previousCharacter.characterComponents + mixingComponents).character
             
             if ligatureCharacter != combinedCharacter && ligatureCharacter.isEmpty == false && combinedCharacter.isEmpty == false {
@@ -356,7 +366,10 @@ class Keyboard: NSObject {
             characterComponents += [.extraDownRight]
         }
         
-        delegate?.delete()
+        if delegate?.documentContext.beforeInput != previousDocumentContextBeforeInput {
+            delegate?.delete()
+        }
+        
         input()
         
         if currentLabel.isEmpty {
